Constructed the hierarchy of source codes
